Revert "[CLANG] Work around weird Clang compiler issue that results in incorrect TPFAIL message due to garbage function return values"
This reverts commit f0097bd6847204514b3f02a428a0f7124041dabb.

At the time of the prior commit, I thought I had found a workaround for the CLang issue.
But turns out the test passed after those code changes because it ran with MM access method
(which the test framework randomly chooses), not because of the code changes. The test failed
again in later testing when the test framework chose BG access method. This clearly shows the
workaround does not work so no point having it. Hence the revert.

This also means that we know of at least one issue that causes incorrect code to be generated when
building YottaDB using CLang. That said, it is only one subtest that fails with PRO builds using
CLang. That is the triggers/trigdefbad subtest.