[#467] SIG-11 from SimpleAPI/SimpleThreadAPI if $ydb_dist/gtmsecshr is not root owned and/or does not have setuid bit set
When $ydb_dist/gtmsecshr is not root owned or does not have setuid bit set, a ydb_init() call core dumps.
Below is the C-stack.

 #0 gtm_free (addr=0x0) at sr_port/gtm_malloc_src.h:887
 #1 stp_gcol () at sr_port/stp_gcol_src.h:1057
 #2 n2s () at sr_port/n2s.c:32
 #3 set_zstatus () at sr_unix/set_zstatus.c:76
 #4 gtmci_ch () at sr_unix/gtmci_ch.c:69
 #5 rts_error_va () at sr_unix/rts_error.c:192
 #6 rts_error_csa () at sr_unix/rts_error.c:99
 #7 ydb_init () at sr_unix/ydb_init.c:282
 #8 main ()

The core dump is because we are trying to issue a GTMSECSHRPERM error but to issue the error we require
the stringpool to be setup which is not yet done (will be done a little later in ydb_init() as part of
calling init_gtm()).

(gdb) f 1
 #1 stp_gcol () at sr_port/stp_gcol_src.h:1057
1057   free(strpool_base);

(gdb) p strpool_base
$1 = (unsigned char *) 0x0

The primary fix is to ydb_init.c to do a stp_init() call before the rts_error_csa() call.

In addition, fixes were done to ydb_init.c to avoid memory bloat if repeated calls to ydb_init()
happen with the same error scenario.

Also, a fix was done to common_startup_init.c to handle the case when it is called multiple times
(as part of multiple ydb_init() calls none of which run to completion due to an error scenario).
Previously it would issue a misleading MIXIMAGE error if a C program calls ydb_init() multiple times
and fails to check the error return status of ydb_init().

Note that the above discussion talks about ydb_init() but any SimpleAPI/SimpleThreadAPI call is also
affected since all of them internally do a ydb_init() as the first step if needed.