[#420] Do not invoke DO_FORK_N_CORE_IN_THIS_THREAD_IF_NEEDED if signalForwarded is TRUE
The DO_FORK_N_CORE_IN_THIS_THREAD_IF_NEEDED macro needs to be executed by the thread that got
the process-terminating signal (e.g. SIGSEGV) to dump a core so we have the C-stack of the SIG-11.
This macro waits for the thread that holds the YDB engine lock to signal it (using a global
variable "safe_to_fork_n_core") to dump a core. Therefore this macro can be executed by a thread
that does not hold the YDB engine lock as long as that is the thread that got the first signal.

It also means this macro should not be executed by any thread that got the signal forwarded (possible
by a pthread_kill() call in ydb_stm_thread() or FORWARD_SIG_TO_MAIN_THREAD_IF_NEEDED macro).
Therefore check for signalForwarded and only if it is FALSE execute this macro.

Also enhance some comments.