[#111] Fix hang in MUPIP RUNDOWN, MUPIP RECOVER or MUPIP ROLLBACK after a kill -9 of process and shared memory is not removed
See https://gitlab.com/YottaDB/DB/YDB/issues/111 for background on the issue.

The real issue is that MUPIP RUNDOWN hangs if the process gets a kill -9 after it has updated
jbp->rsrv_freeaddr but before it has updated jbp->phase2_commit_index2 (the CMT06 step in the
commit logic). Cases like these are normally handled by the JNL_PHASE2_CLEANUP_IF_POSSIBLE
macro call in jnl_write_attempt(). But that works only if jbp->phase2_commit_index1 !=
jbp->phase2_commit_index2. Which is not helpful for us in this kill -9 scenario since it is possible
for phase2_commit_index1 to be == phase2_commit_index2.  Therefore, we now add an extra check in
mu_rndwn_file() to handle the kill -9 in CMT06.

Coincidentally, GTM-8989 in GT.M V6.3-005 had addressed a similar issue with the below release note
at http://tinco.pair.com/bhaskar/gtm/doc/articles/GTM_V6.3-005_Release_Notes.html#GTM-8989

```
MUPIP RUNDOWN works as expected. In rare situations, processes killed during transaction commits
could leave the transaction information in an inconsistent state that caused MUPIP RUNDOWN to
hang. Note FIS strongly recommends against kill -9 of a process accessing the database as it can
induce database damage. This issue was only observed in the GT.M development environment, and was
never reported by a user. (GTM-8989)
```

But the fix is incomplete (we have a test case that demonstrates the issue after some white-box code
changes in a prior commit as part of https://gitlab.com/YottaDB/DB/YDB/merge_requests/637). This is
because it did the INCR_PHASE2_COMMIT_INDEX macro invocation (which addresses the issue by ensuring
index != index2 when we invoke JNL_PHASE2_CLEANUP_IF_POSSIBLE in jnl_write_attempt()) only if
index1 == index2. But if the process did a lot of updates before it got kill -9ed, we would have
a sequence of phase2 commits that have not yet been cleaned up by jnl_phase2_cleanup() followed
by a phase2 commit that has not yet been salvaged by jnl_phase2_salvage().  In that scenario, we
will not invoke INCR_PHASE2_COMMIT_INDEX in mu_rndwn_file(). Later calls to jnl_write_attempt()
(invoked by wcs_flu() -> jnl_flush()) would invoke jnl_phase2_cleanup() on the updates that can be
cleaned up. Eventually it would land on the update that needs to be salvaged (because the process
got kill -9ed) but at that point index1 could be == index2 and so the JNL_PHASE2_CLEANUP_IF_POSSIBLE
macro would do nothing resulting in a hang/loop between jnl_flush() and jnl_write_attempt().

This is now addressed by not checking if index1 == index2 in mu_rndwn_file(). And a lot of the checks
that were added in V6.3-005 by GTM-8989 are now removed. Instead HANDLE_KILL9_IN_CMT06_STEP_IF_NEEDED
is a new macro that is used here. This is modeled on code that already existed in the
UPDATE_JBP_RSRV_FREEADDR macro (added as part of GT.M V6.3-006 but could not locate any release note
mention of this change) and that previous code also has been enhanced to use the new macro.