Enhance MUPIP RUNDOWN to handle kill -9 in the middle of journal file switch
We had the ideminter_rolrec/idemp_rollback_or_recover subtest fail twice in the past
3 months with the following failure symptom.

%YDB-F-ASSERT, Assert failed in sr_unix/wcs_flu.c line 327 for expression (jb->fsync_dskaddr == jb->freeaddr)

(gdb) where
 #0  __pthread_kill () at ../sysdeps/unix/sysv/linux/pthread_kill.c:56
 #1  in gtm_dump_core () at sr_unix/gtm_dump_core.c:72
 #2  in gtm_fork_n_core () at sr_unix/gtm_fork_n_core.c:163
 #3  in ch_cond_core () at sr_unix/ch_cond_core.c:80
 #4  in rts_error_va () at sr_unix/rts_error.c:194
 #5  in rts_error_csa () at sr_unix/rts_error.c:101
 #6  in wcs_flu () at sr_unix/wcs_flu.c:327
 #7  in mur_open_files () at sr_port/mur_open_files.c:537
 #8  in mupip_recover () at sr_port/mupip_recover.c:195
 #9  in mupip_main () at sr_unix/mupip_main.c:124
 #10 in dlopen_libyottadb () at sr_unix/dlopen_libyottadb.c:148
 #11 in main () at sr_unix/mupip.c:19

(gdb) f 6
 #6  in wcs_flu () at sr_unix/wcs_flu.c:327
 327                             assert(jb->fsync_dskaddr == jb->freeaddr);

   325   if (jb->last_eof_written)
   326   {
-> 327           assert(jb->fsync_dskaddr == jb->freeaddr);
   328           assert(jb->rsrv_freeaddr == jb->freeaddr);
   329           assert((dba_bg != csd->acc_meth) || !csd->jnl_before_image

(gdb) p/x jb->rsrv_freeaddr
$2 = 0x214c838

(gdb) p/x jb->fsync_dskaddr
$3 = 0x214c758

(gdb) p/x jb->freeaddr
$4 = 0x214c838

This is a case of a kill -9 that happened after jb->last_eof_written was set to TRUE in jnl_file_close()
but before the jnl_fsync() finished.

The fix for this case is to invoke jnl_fsync() again as part of the MUPIP RUNDOWN/RECOVER/ROLLBACK
to set jb->fsync_dskaddr back to the correct value.

In pro builds, it is not clear that there is any user-visible issue. Hence no separate issue created
on gitlab for this fix.