[#467] Avoid duplicate stp_init() call if going through ydb_init(); Ensure enough space in stringpool before initializing $SYSTEM
If YottaDB is started through ydb_init() (e.g. in a C program), then stp_init() is done even
before init_gtm() or gtm_startup() happens (due to a prior #467 commit). This means when a
later call to stp_init() hapens inside gtm_startup() we would effectively create a new
stringpool and lose track of the previous one i.e. a small (~ 100K) memory leak. This is now
fixed by freeing up the prior stringpool before allocating the second one if this is indeed
the second call to stp_init() at YDB init time. This is mostly a cosmetic fix.

The more important change is to dollar_system_init.c. The v62002/gtm8167 subtest failed occasionally
on only one box in our in-house testing with the following diff.

> v62002_0_12/gtm8167/gtm8167.mje
> %YDB-F-ASSERT, Assert failed in sr_port/dollar_system_init.c line 55 for expression (stringpool.free < stringpool.top)

The assert failed because we assumed that in dollar_system_init(), the stringpool would be almost empty
because this is invoked at process startup and so we do not do any checks of whether the stringpool has
space to copy over the $SYSTEM ISV. Instead we have an assert that there is indeed space to do the copy.
And it is this assert that failed. This is because if the process starts up as the child of a JOB command,
it records parameters passed in the stringpool (C-stack : gtm_startup -> jobchild_init -> ojchildparms).
And this can cause the stringpool to be almost full when we later come to dollar_system_init(), particularly
if the gtm_sysid env var is as long as the maximum of 256 bytes. The failure happens only some times
because in stp_init(), we invoke malloc() with 2 os page size padding and do an os page size alignment of the
return pointer value.  Depending on the actual value returned by malloc(), we might end up with a little more
or a lot more than the requested amount and the failure happens only when we end up with only a little more
space left after filling in the job parameters (which are 32 of them each 8K in size in this gtm8167 subtest).
The fix is straightforward and is to ensure enough stringpool space is available before doing the copy.
While it is clear that this is an issue in the dbg build (I was able to get a failure even with GT.M V63005),
it is not yet clear what the implications of this are in pro particularly because $gtm_sysid is usually not
set by any user. Therefore no separate user-visible issue is created for this fix.