[DEBUG-ONLY] Enhance jb->blocked assert in jnl_write.c to account for defunct processes
If the process id corresponding to jb->blocked was abnormally killed (e.g. kill -9), it is possible
the process is still alive but in a defunct state (when another process comes into jnl_write.c
after having salvaged crit etc.) if the parent process has not yet done a waitpid() on this abnormally
terminated child. In that case, the following assert would fail.

    266     assert((!jb->blocked) || (FALSE == is_proc_alive(jb->blocked, 0)));

This is now addressed by using the new is_defunct_pid() function in the assert.