[#472] ydb_data_s() and ydb_data_st() should not return 0 on an error (with [YottaDB/Lang/YDBGo#6])
We define the expected return codes now in libyottadb.h as YDB_DATA_*. One of those codes is
YDB_DATA_ERROR which is returned in *ret_value when ydb_data_s() or ydb_data_st() get an error
and return a non-YDB_OK return code.

We do not really recommend that users rely on YDB_DATA_ERROR because the recommended coding
method is to check for error values in where they are returned before trying to look
at the return value. But for those who wish it, we are now providing a code that is not a
valid data-describing return value.

Previously these functions were returning 0 in this error condition which
itself is a valid return value.