Revert "[#410] Allow user to define which signals will be forwarded to non M main's signal handlers"
This reverts commit 4b26b88c43668474dee53e6da4c2132beb81dae6.

With this update, some regression tests with versions of Go past 1.10 failed with destroyed
stacks, which did not fail with this update. Although no cause-effect relationship could
reasonably be established, the correlation between passing and failing tests and this update
means that this update is held off, at least for the present. So it is reverted for now.