[#478] Add SA_ONSTACK and SA_SIGINFO flags to all signal handlers, fix altstack size issue with Go main routines
The SA_ONSTACK option is required for signal handlers created in a Go environment. The SA_SIGINFO option is not
required for that but is required for forwarding but the two were so interlinked in [#410] it was more difficult
to separate them so the SA_SIGINFO changes (eventually needed anyway) came along for the ride.

The SA_ONSTACK issue was causing problems in the rundown_argless/ipcsnoreset test and others. This test sends a
SIGTERM (MUPIP STOP) to each of the impjobgo processes when the Go version of imptp is selected (imptp flavor 3).
When a signal comes in and is caught by a handler with the SA_ONSTACK flag set, execution of the handler is moved
to an alternate stack that was setup by Go. This alternate stack was only 32KB but one of the routines that runs
during rundown is jnl_file_close() which has a stack variable that is 67KB which caused the stack to overrun
resulting in a sig-11. This change checks if an alternate stack is allocated and if so, sees if it is at least
128KB and if not, allocates and sets a new stack (old stack is reinstalled when YDB is closed down).