[#480] Issue with ydb_incr_s() of non-existant global with literal value giving incorrect results
This issue was first seen with the SimpleAPI but it exists with the M interface as well. A simple
M program shows the issue:
	set ^test=0
	set x="1000000" ; Create string 7 digit value
	set y=$incr(^test,x)
	zwrite
	quit
Without the fix, y shows up in the zwrite as the null string.

If the value is 6 digits or fewer, it is an integer - else it is a floating point. The two numeric
types go through different paths. In the integer version, no check is done on the values - they are
added together and checked for overflow. In the floating version, things are complex enough there
is a check to see if the original value is 0 and if so, just returns the increment mval as the
result. This is what causes us to return a broken mval for 7 digits or more. So there are two
pieces to this issue:
  1. The incremented value must be 0 or undefined.
  2. The increment must be 7 or more digits.