Revert "Fixed hundreds of maybe uninitialized compiler warnings"
This reverts commit a92f3f31e1c3f44e095245abba6a1a15b80c2de5.

This commit is being reverted because it caused problems in 4 subtests:
1. triggers/gtm7877
2. triggers/gtm7877a
3. collation/def_to_pol
4. collation/def_to_chn
Additional considerations are that most of these changes should be unnecessary if
the maybe-uninitialized warnings gcc generates were trust worthy (they are not as
the support has many bugs listed against it). In a re-review, all of the items
looked at were either irrelevant or just simply incorrect.