[#496] Fix M object code generation for loader (ld) version 2.33 or later Changes in the loader were picked up that are enforcing (and/or using) additional flags and values in the elf headers that were not used before which is causing errors trying to load YDB object files to create shared libraries. This is even causing installs and/or builds to fail.
As noted in the issue [#496] this is not a 100% fix because eu-elflint still
calls out an error but it is not one the loader/linker has issues with - at
least not yet.

For now this is only an issue on Ubuntu 19.10 and the latest Arch as far as is
known but it will eventually cause issues on subsequent releases in all distros
as they pick up these changes too.