[#496] Generate correct ELF object code wrapper for M object code.
The earlier fix for this issue was noted as incomplete as there was still an error from eu-elflint run
against a generated M object file. But this remaining issue also broke the ability of an M routine to be
be run out of a shared library. We discovered this when gitlab pipelines and some tests started failing
when they tried to run ydb_set_env which drove ^%YDBENV. The error was that %YDBENV could not be found.
It turned out that while the loader accepted YDB object files and put them in the shared library, the
output symbols were not able to be located by dlsym() meaning the shared library was useless.

Further investigation showed that the symbol table in the ELF object file that encapsulates the M
object, had entries that were not in the correct order. ELF requires that the local symbols come
BEFORE the global symbols but we were generating the symbols in the order LCL, GBL, LCL so not matter
what value was filled in for sh_info (which is supposed to contain the highest LOCAL symbol index
plus one) we got an error. Swapping the order of the 2nd and 3rd symbols being generated in the
symbol table cured this issue such that both eu-elflint says no errors found and ydb_env_set now runs
without any issues. In this new configuration, shdr->sh_info is set to 2 since the local symbols are
at indexes 0 and 1.