[DEBUG-ONLY #500] Prevent assert in error_return() when simpleapi call drives trigger that gets an error
A case was recently encountered during Octo development where a simpleapi ydb_set_s() call was setting a
global that drove a trigger. The trigger invoked a non-existent routine which generated an error. But
gtm_trigger() had set SFF_NORET_VIA_MUMTSTART_OFF in the frame prior to the trigger base frame (which in
this case was a call-in base frame) but with the error, gtm_trigger_fini() was not turning the flag off.
This caused an assert in error_return() line 146. The fix is to always remove the SFF_NORET_VIA_MUMTSTART_OFF
flag when unwinding the trigger frame as there's nothing else to do it in this situation as
mdb_condition_handler is no longer in the picture with the dm_start frame already unwound.