[#496] Fix issue with generated M code ELF object wrapper with non-null terminated string table
After the previous two fixes, latest issue to show itself with the ELF wrapper for generated
M object code is the string table was not properly null terminated (which was mentioned by the
gold loader [ld.lld] whereas the regular loader just gave a "bad value" error.

There were two issues here - (1) the last element in the string table (the mident_fixed used
to hold the routine name) had a null terminator but the last character of the block also
needed to be null terminated. So changes were made to zero out the unused section of the mident.
(2) The other issue was too much storage was being allocated for the the string table. The
calculation was using (SIZEOF("literal") + 1) when the SIZEOF() the literal already held room
for the null. Removing the extraneous "+ 1"s made things work again. Also added an assert that
the mident_fixed was the last thing in the string table so its null termination served as the
null termination for the string table itself.

Also changed the language name from "MUMPS" to "M".