[#504] Address issues when an error is raised during a commit on a 32 bit platform (affects armv6l and armv7l)
Issue was an array of 64 bit values was being passed but was not being handled correctly in the secshr_send_DBCLNUPINFO_msg.c routine. This issue started occurring when the build was changed from an LTO build to a non-LTO build for armv6l system which we did because armv6l was no longer able to build with LTO (lots of linker errors from gold loader complaining about something missing).

Also address minor issues with printing addresses in various messages and debug outputs across 32 and 64 bit platforms (using the correct format codes) in several modules - most are debugging messages with enabling flags around them. Minor additions also so some of the debugging mode flags build correctly as well. Also some minor cleanups such as there was no need for DBGTRIGR macros to be separately defined in op_fntext.c for example.