Make function prototypes `const` where possible
This is 100% backwards compatible with the existing C API,
but makes the Rust API much easier to use.

In Rust, mutable pointers are never allowed to alias,
and having two mutable pointers to the same data is undefined behavior.
As a result, taking a mutable reference forms an exclusive lock checked at
compile time by the 'borrow checker'. Changing an API to take an
immutable reference makes it much easier to use for the caller.

The reason these changes had to be made in C instead of in the wrapper is two-fold:

1. First, mutating an immutable pointer is undefined behavior in Rust, so
casting a `*const` to `*mut` is unsound, and
2. Changing the generated bindings not only is difficult to maintain,
but gives no guarantee that the C code is not mutating the pointer.