[#515] ydb_zstatus() should not blindly copy into a buffer passed with a zero length
Since the ydb_zstatus() entry point is exported, this is an issue both with users calling it
and with it being called internally - specifically by ydb_simpleapi_ch() when a threaded
simple API call for which the user code had passed in an unallocated errstr (ydb_buffer_t)
and got an error. This ended in an assert failure/sig-11 depending on build.

Additionally, ydb_zstatus now has a return code that code can check to verify it got the
entire message. If insufficient room is provided, the YDB_ERR_INVSTRLEN return code is
returned else, the code is YDB_OK.