Revert "Significant changes to make YottaDB quick start (https://yottadb.com/product/get-started/) language agnostic:"
This reverts commit b65ea76825c194d0d38ed2122b441b03f77b89d1.

That commit changed the `yottadb-base` docker image to start in UTF-8 mode which caused pipelines
of Octo (and maybe other repos that rely on YDB repo) to fail as they were expecting M mode.

The original intent of that commit can be satisfied by creating a new `debian` docker image of
`yottadb` just like a `centos` and `ubuntu` images are already created. This will be done in a separate
commit taking care not to affect `yottadb-base` docker images.