[#390] Fix sig-11 in $ZYHASH on ARM due to unaligned accesses
In our automated testing, the r130/ydb390 test has occassionally sig-11ed on 32 bit ARM machines. The sig-11 happened in the MurmurHash3_x64_128() call when it attempted to access a valid subscript of string->str.addr. This failure was likely to due to an unaligned access that isn't fully supported on ARM.  This change fixes this failure by defining UNALIGNED_ACCESS_FULLY_SUPPORTED when the architecture fully supports unaligned access but not on 32 bit ARM machines.