[#485] Cleanups plus zero suppression fix for #485
This change adds a debugging macro (DBG_VALIDATE_MVAL) to mdef.h so is available everywhere.
This macro allows a created mval to be validated. Current validations are limited to those
in the comments but over time these validations could be expanded as can the use of this macro.

This change also fixes the issues seen with #485 in its original form. There were several small
nits missed in the final review as well as the larger issue that caused test failures when the
decimal value '0' was converted to hex. It resulted either in a null string or, if the stringpool
where the value was places was full of ascii '0' characters, it generated large negative lengths
for the mval string leading to asserts and/or sig-11 exceptions depending on the build type
(dbg or pro).