[#520] Fixed bug where repeatedly switching between $ETRAP and $ZTRAP could cause assert failures in $ZTRAP
The issue was caused by the mval being placed within the indirect frame, which is sometimes shifted as part of the process for cleaning up $ETRAP values when setting a $ZTRAP value. The fix involves modifying gtm_newintrinsic() to make it return an integer (the number of bytes shifted) so that the mval can be shifted the same number of bytes.