[#524] Automatically convert endianness of global directories to the correct endianness
This commit modifies GDE to allow global directories created with a version of the upstream software running on a big endian system to be converted to a YottaDB version running on a little endian system.

The many coding standards issues with this M file were left alone as changing the whole file to spell out the full names of M commands would make it extremely difficult to figure out what had been changed by looking at a diff.