[DEBUG-only] Allow process to already have crit when it grabs it if process is already shutting down
This is a situation that can occur if a fatal signal comes in after we grabbed crit for something but
before we released it. Cleanup processing will also try to grab crit and in a debug build, will note
that we already have crit and assert fail. This debug-only change allows the assert to succeed if
the process is in exit processing and we just return as if we had acquired the lock. This means the
lock will be released by the cleanup code but this is not a problem because we won't be returning to
the code that originally grabbed the lock in the first place allowing process exit processing to
complete.