[#532] Fix issues with installing encryption plugin in UTF-8 mode
Previously, the encryption plugin would not install in UTF-8 mode because its makefile was not passing $ydb_icu_version to pinentry.m when it attempted to compile it in UTF-8 mode. This bug has been fixed. In addition, ydb_env_set now automatically detects and sets the ICU version if it is not set.