[#521] Added --octo option to install script and fixed a few install script bugs
The major change in this commit is the addition of the --octo option, which builds the YDBOcto plugin from source automatically. As there is currently no release build for Octo, this builds the latest master.

This commit also addresses issues regarding the zlib plugin. Previously, it would not install in UTF-8 mode if default was specified as the ICU version. This has been fixed. In addition, the name of the repo has been updated to YDBZlib.

Finally, there is another minor change involving the UTF-8 flag. Previously, it was not set if UTF-8 mode was not specified which resulted in an error message when the POSIX plugin checked if the flag was set to "Y". This has been fixed.