[DEBUG-ONLY] $ZPEEK() with a NULL command can assert fail due to references to garbage memory
This issue occurred as an assert failure hence the DEBUG-ONLY designation. There is a very SLIGHT chance
that this issue could fail in a production version but it would take some effort and is all but impossible
to do accidently. The fix is in namelook() adding protection against a specified zero length string and
preventing references to the first byte of the non-existent buffer (and an assert that fails).