[DEBUG-ONLY] Remove assert about dead prior crit owner in jnl_write.c as it is not valid in EOWNERDEAD case
The below assert failed twice in the past few months of in-house testing (both times in
the `ideminter_rolrec` test) where `kill -9` of processes are involved.

```c
assert(!blocking_pid || is_defunct_pid((uint4)blocking_pid) || (FALSE == is_proc_alive(blocking_pid, 0)));
```

This indicates that even though the current process (say P1) has seized the critical section lock from what
was a killed process (blocking_pid, let us say P2), P1 finds that P2 still is alive.

The man pages of pthread_mutex_trylock() say the following.

https://pubs.opengroup.org/onlinepubs/9699919799/functions/pthread_mutex_lock.html

```
If mutex is a robust mutex and the owning thread terminated while holding the mutex lock, a call to
pthread_mutex_lock() may return the error value [EOWNERDEAD] even if the process in which the owning
thread resides has not terminated.
```

Therefore in the EOWNERDEAD case, if `mutex_clean_dead_owner()` is called in `sr_unix/mutex.c`, and
the critical section lock is salvaged, we are not guaranteed the prior owning pid is dead and so
the above mentioned assert in `sr_port/jnl_write.c` has to be now removed.

The assert is currently the only user of the `is_defunct_pid()` function and so that function can
also be removed from the code base. But since it is a utility function and is expected to be needed
for some other purpose in the near future, it is kept as is.