[#553] Compiling boolean expressions involving NOT operator (') and extrinsic function calls spin loops if ydb_boolean=1
The issue was in `sr_port/bx_boolop.c` in the logic that determined if the current triple is followed
by a `OC_BOOLFINI` triple. In case the NOT operator (`'`) is used, the `OC_BOOLFINI` would follow the
current triple but after an intermediate `OC_COM`. This caused `bx_boolop.c` to not properly move
the boolean triples into the boolchain (`TREF(boolchain_ptr)`) resulting in assert failures (in dbg builds)
and infinite loops (in pro builds).

Note that it is possible a boolean expressions has an arbitrary sequence of the NOT operator and so
we need to check for not just one but any number of `OC_COM` triples before we see the `OC_BOOLFINI`.