[#556] Add ydb_ci_get_info()/ydb_ci_get_info_t() entry points to support call-ins from wrappers (fetches parameter info for routines)
When a call-in is made, YottaDB reads a call-in table created by the user which defines
which parameters are input-only, input-output, or output-only. The wrapper also needs
this information so it can copy an output parm from C storage back to Go storage. If it
doesn't know which parms are output, it cannot do this. So these routines are added that
currently just give the input and output parameter bit masks (used together they identify
the 3 catagories mentioned). This allows the wrapper (with changes made in YDBGo#28) to
properly handle output parameters.

The current state of these routines is temporary as future releases will expand on this to
also return the type of the parameters so we can do some error checking and also support
output and return types other than string.