Fix rare possibility of assert failure in `assert(GET_DEFERRED_EXIT_CHECK_NEEDED || (1 != forced_exit));`
In-house testing revealed a rare test failure in the v53003/D9I10002706 subtest.
A MUPIP REORG process started by this subtest was sent a MUPIP STOP by the subtest and assert failed.

```
> cat ./mupip_reorg_7.logx
Fill Factor:: Index blocks 100%: Data blocks 100%
%YDB-F-ASSERT, Assert failed in sr_port/db_csh_getn.c line 480 for expression (GET_DEFERRED_EXIT_CHECK_NEEDED || (1 != forced_exit))

(gdb) where
 #0  __pthread_kill () at ../sysdeps/unix/sysv/linux/pthread_kill.c:56
 #1  gtm_dump_core () at sr_unix/gtm_dump_core.c:74
 #2  gtm_fork_n_core () at sr_unix/gtm_fork_n_core.c:163
 #3  ch_cond_core () at sr_unix/ch_cond_core.c:80
 #4  rts_error_va () at sr_unix/rts_error.c:192
 #5  rts_error_csa () at sr_unix/rts_error.c:99
 #6  db_csh_getn () at sr_port/db_csh_getn.c:480
 #7  t_qread () at sr_port/t_qread.c:444
 #8  gvcst_search () at sr_port/gvcst_search.c:526
 #9  gvcst_data2 () at sr_port/gvcst_data.c:150
 #10 gvcst_data () at sr_port/gvcst_data.c:76
 #11 op_gvdata () at sr_port/op_gvdata.c:48
 #12 gv_select_reg () at sr_port/gv_select.c:373
 #13 gv_select () at sr_port/gv_select.c:302
 #14 mupip_reorg () at sr_port/mupip_reorg.c:238
 #15 mupip_main () at sr_unix/mupip_main.c:122
 #16 dlopen_libyottadb () at sr_unix/dlopen_libyottadb.c:148
 #17 main () at sr_unix/mupip.c:22

(gdb) f 6
 #6  0x00007fd5b9bd69da in db_csh_getn (block=1343) at sr_port/db_csh_getn.c:480
480                     ENABLE_INTERRUPTS(INTRPT_IN_DB_CSH_GETN, prev_intrpt_state);

(gdb) p forced_exit
$4 = 1

(gdb) p deferred_signal_handling_needed
$5 = 4
```

The assert that failed evaluates to TRUE in the core. That is `GET_DEFERRED_EXIT_CHECK_NEEDED` is TRUE
and `forced_exit` is 1. I suspect what happened is that after `GET_DEFERRED_EXIT_CHECK_NEEDED` macro got
executed, the MUPIP STOP was received thereby interrupting the assert and then going to execute
`generic_signal_handler()` which then set `forced_exit` to 1 as well as made `GET_DEFERRED_EXIT_CHECK_NEEDED`
macro TRUE before returning to the assert code that then checked `forced_exit` was 1 and failed the assert.

To prevent failures like this all we need is to reverse the operands in the `||` condition. This way
we check `1 != forced_exit` first. And only if that is not TRUE (i.e. `1 == forced_exit`) will we even
go to do the `GET_DEFERRED_EXIT_CHECK_NEEDED` check. At that point, we are guaranteed the macro will
evaluate to a non-zero value.