Enhance ydbinstall.sh to skip releases that do not yet have binary tarballs before determining latest official release
A user tried using `ydbinstall.sh` when `r1.30` release had been created but it was still in draft mode
(i.e. no binary tarballs had been uploaded yet but release notes were up and under construction).
The `ydbinstall.sh` invocation failed trying to download the `r1.30` tarballs. In this case, it should
have skipped `r1.30` (because it is not yet an official release) and moved on to `r1.28` as the latest
official release (it has binary tarballs) and tried downloading/installing that release.

This is now fixed by examining the downloaded `tags` file from gitlab and looking for releases with
binary tarballs by grepping for ".pro.tgz" and only picking such releases as candidates before
determining the latest release.