[#563] Enhance ydbinstall.sh to make CentOS 8.x a supported platform
As we now offer a binary tarball for CentOS 8.x, ydbinstall.sh needed to be updated so that the binary tarball can be installed on CentOS 8.x without using the --force-install option. This commit makes the necessary change. This commit also updates the script to download the CentOS 8 tarball instead of the RHEL 7 tarball on CentOS/RHEL/SLES 8 when the install script is run without an already built YottaDB version.