[#576] $PIECE in a database trigger returns incorrect results when invoked from SimpleAPI (e.g. ydb_set_s)
Issue description
------------------
This is an issue that was noticed Octo (see YottaDB/DBMS/YDBOcto!583 (comment 333919051) for details
on the failed pipeline jobs). Octo uses the SimpleAPI to do database updates and has triggers installed
for those updates. A $PIECE invocation inside the M trigger code returned incorrect results. Below
are the details of that particular failure.

The M code that was invoked was the following.

```m
SET newValue=$PIECE($ZTVALUE,$ZTDELIM,$ZTUPDATE)
```

And below is the value of the involved variables at the time of the above $PIECE invocation.

```m
$ZTVALUE="LOTSOFCOLS|2200|16388|0|16385|0|16386|0|0|0|0|16389|1|0|p|r|3|0|0|1|0|0|0|0|0|1|d|0|571|1||||803051"
$ZTDELIM="|"
$ZTUPDATE=1
```

We therefore expect newValue to be set to LOTSOFCOLS. But instead it is set to LOTSOF (this in turn
results in an incorrect update to the ^%ydboctoxref global and in turn results in an incorrect output
of a pg catalog query and the TC013 subtest validation routine catches this and fails the test).

This failure is seen only in Release builds. Never in Debug builds. Even with Release builds, this
failure happens only rarely and only on some servers so it is most likely an uninitialized variable
issue in the YDB code base.

Fix description
----------------
The issue turned out to be an uninitialized `fnpc_indx` field of the `mval` structure that `ydb_set_s()`
passes as the value to set in the global variable node. This could end up having an arbitrary value and
if it happens to be a non-zero value, then $PIECE could return incorrect results.

This failure is not easy to reproduce since it requires the uninitialized value to be set to a non-zero
value which seems to be an exception rather than the rule (we did not see this in internal testing on
lots of platforms). Therefore it is not possible to come up with a simple automated test for this issue
other than the Octo test case that detected this issue. If that Octo test case passes with this fix, it is
considered enough confirmation of this code fix.